requirements.txt
click==8.1.7
rich==13.7.0
python-dateutil==2.8.2